+++ rtmlib/include/geometry/intersections.h
@@ -11,7 +11,7 @@ namespace rtm
 class Intersections
 {
 public:
-    virtual float get_t() const = 0;
+    virtual double get_t() const = 0;
     virtual Surface* get_shape() const = 0;
 };
 } //namespace rtm
