+++ rtmlib/include/mat_entities/vector.h
@@ -10,24 +10,24 @@ class Vector : public Tuple
 {
     public:
         Vector();
-        Vector(const float& x, const float& y, const float& z);
-        Vector(const float& x, const float& y, const float& z, const float& w);
-        Vector(float* elements);
+        Vector(const double& x, const double& y, const double& z);
+        Vector(const double& x, const double& y, const double& z, const double& w);
+        Vector(double* elements);
 
-        inline float x() const {return this->t[0];}
-        inline float y() const {return this->t[1];}
-        inline float z() const {return this->t[2];}
-        inline float w() const {return this->t[3];}
+        inline double x() const {return this->t[0];}
+        inline double y() const {return this->t[1];}
+        inline double z() const {return this->t[2];}
+        inline double w() const {return this->t[3];}
 
-        inline void set_w(const float& w_val) { t[3] = w_val; }
+        inline void set_w(const double& w_val) { t[3] = w_val; }
 
         std::string to_str() const override;
 
         Vector operator- () const;
 
-        float magnitude() const;
+        double magnitude() const;
         Vector normalize() const;
-        float dot(const Vector&) const;
+        double dot(const Vector&) const;
 };
 } //namespace rtm
